+++ src/components/rssFeed.jsx
@@ -3,10 +3,10 @@ import { makeStyles } from "@material-ui/core";
 import CategorySelector from "./categorySelector";
 import RssEntryList from "./rssEntryList";
 import { fetchCategoryFeeds } from "../libs/parser";
-import feeds from "../../assets/feeds";
+import { feedCollections } from "../../assets/feeds";
 import { rssStyles as styles } from "./styles";
 
-const categoryList = Object.keys(feeds);
+const categoryList = feedCollections.map((item) => item.name);
 const useStyles = makeStyles(() => styles);
 const RssFeed = () => {
   const [categories, setCategories] = useState([]);