+++ src/libs/parser.js
@@ -1,17 +1,20 @@
 import { parseString } from "xml2js";
-import feeds from "../../assets/feeds";
+import { feedCollections } from "../../assets/feeds";
 
-const corsProxy = "https://cors.eu.org/";
-
+const urlRssYouTube = "https://www.youtube.com/feeds/videos.xml?channel_id=";
 const fetchCategoryFeeds = (category, callback) => {
-  const { length } = feeds[category];
+  const index = feedCollections.findIndex((item) => item.name === category);
+  if (index === -1) {
+    return;
+  }
+  const { length } = feedCollections[index].feeds;
   for (let idx = 0; idx < length; idx++) {
-    fetchFeed(feeds[category][idx], callback);
+    fetchFeed(feedCollections[index].feeds[idx], callback);
   }
 };
 
-const fetchFeed = (url, callback) => {
-  fetch(corsProxy + url)
+const fetchFeed = (item, callback) => {
+  fetch(urlRssYouTube + item.channelId)
     .then((res) =>
       res
         .text()
@@ -34,7 +37,6 @@ const parseStringCallback = (err, res, callback) => {
     const entries = res.feed.entry.map((item) => {
       return cleanParsedXMLEntry(item);
     });
-    console.info("entries", entries);
     callback(entries);
   }
 };